Renaming and formatting of POI example
