+++ Examples/VisionExamples/Info.plist
@@ -22,6 +22,16 @@
 	<true/>
 	<key>MGLMapboxAccessToken</key>
 	<string></string>
+	<key>NSCameraUsageDescription</key>
+	<string>Frames from the camera are passed to ML models to get semantic segmentation, object detection and extract other features</string>
+	<key>NSLocationAlwaysAndWhenInUseUsageDescription</key>
+	<string>Location and heading information is used for position estimation, localization and AR-related tasks</string>
+	<key>NSLocationAlwaysUsageDescription</key>
+	<string>Location and heading information is used for position estimation, localization and AR-related tasks</string>
+	<key>NSLocationWhenInUseUsageDescription</key>
+	<string>Location and heading information is used for position estimation, localization and AR-related tasks</string>
+	<key>UIFileSharingEnabled</key>
+	<true/>
 	<key>UILaunchStoryboardName</key>
 	<string>LaunchScreen</string>
 	<key>UIMainStoryboardFile</key>
@@ -42,15 +52,5 @@
 		<string>UIInterfaceOrientationLandscapeLeft</string>
 		<string>UIInterfaceOrientationLandscapeRight</string>
 	</array>
-    <key>MGLMapboxAccessToken</key>
-    <string></string>
-    <key>NSCameraUsageDescription</key>
-    <string>Frames from the camera are passed to ML models to get semantic segmentation, object detection and extract other features</string>
-    <key>NSLocationAlwaysAndWhenInUseUsageDescription</key>
-    <string>Location and heading information is used for position estimation, localization and AR-related tasks</string>
-    <key>NSLocationAlwaysUsageDescription</key>
-    <string>Location and heading information is used for position estimation, localization and AR-related tasks</string>
-    <key>NSLocationWhenInUseUsageDescription</key>
-    <string>Location and heading information is used for position estimation, localization and AR-related tasks</string>
 </dict>
 </plist>
